+++ gui/headers/TriangleSelectionSmallestInternalAngle.h
@@ -6,7 +6,7 @@
 class TriangleSelectionSmallestInternalAngle: public TriangleSelection{
 public:
     TriangleSelectionSmallestInternalAngle();
-    Triangle* process(Mesh* mesh, int value);
+    Triangle* process(Mesh* mesh, double value);
     ~TriangleSelectionSmallestInternalAngle();
 };
